Separates function declarations and refactors arithmetic operators.
